requirements.txt
absl-py==2.0.0
accelerate==0.29.2
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.13.1
antlr4-python3-runtime==4.8
anykeystore==0.2
apex==0.1
appdirs==1.4.4
archspec @ file:///home/conda/feedstock_root/build_artifacts/archspec_1708969572489/work
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
automodel==0.5.9
backcall==0.2.0
banal==1.0.6
biopython==1.83
biotite==0.40.0
bitarray==2.7.6
bitsandbytes==0.43.1
blessed==1.20.0
blinker==1.6.3
boltons @ file:///home/conda/feedstock_root/build_artifacts/boltons_1711936407380/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1695989787169/work
cachetools==5.3.2
certifi==2023.5.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
click==8.1.7
cmake==3.29.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
conda @ file:///home/conda/feedstock_root/build_artifacts/conda_1711445837045/work
conda-libmamba-solver @ file:///home/conda/feedstock_root/build_artifacts/conda-libmamba-solver_1706566000184/work/src
conda-package-handling @ file:///home/conda/feedstock_root/build_artifacts/conda-package-handling_1691048088238/work
conda_package_streaming @ file:///home/conda/feedstock_root/build_artifacts/conda-package-streaming_1691009212940/work
contextlib2==21.6.0
contiguous-params==1.0.0
contourpy==1.1.1
cryptacular==1.6.2
cycler==0.11.0
Cython==3.0.0
dataset==1.6.2
datasets==2.15.0
decorator==5.1.1
deepspeed==0.5.9
defusedxml==0.7.1
dill==0.3.7
distro @ file:///home/conda/feedstock_root/build_artifacts/distro_1704321475663/work
DLLogger @ git+https://github.com/NVIDIA/dllogger.git@0540a43971f4a8a16693a9de9de73c1072020769
dm-tree==0.1.8
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
dpcpp-cpp-rt==2024.0.2
einops==0.7.0
et-xmlfile==1.1.0
eval_type_backport==0.2.0
evaluate==0.4.1
executing==2.0.1
fair-esm @ git+https://github.com/facebookresearch/esm.git@2b369911bb5b4b0dda914521b9475cad1656b2ac
fairseq==0.12.2
filelock==3.12.2
Flask==3.0.0
flit_core==3.9.0
fonttools==4.25.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.11
GitPython==3.1.43
goatools==1.3.11
google-auth==2.25.2
google-auth-oauthlib==1.0.0
gpustat==1.1.1
greenlet==3.0.3
grpcio==1.60.0
hjson==3.1.0
huggingface-hub==0.23.3
hupper==1.12
hydra-core==1.0.7
idna==3.4
importlib-metadata==7.0.1
importlib-resources==6.0.0
intel-cmplr-lib-rt==2024.0.2
intel-cmplr-lic-rt==2024.0.2
intel-opencl-rt==2024.0.2
intel-openmp==2024.0.2
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.1
jsonpatch @ file:///home/conda/feedstock_root/build_artifacts/jsonpatch_1695536281965/work
jsonpointer @ file:///home/conda/feedstock_root/build_artifacts/jsonpointer_1695397259927/work
kiwisolver==1.4.4
libmambapy @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1711394305528/work/libmambapy
lightning-utilities==0.9.0
lightseq==3.0.1
lit==18.1.7
lxml==4.9.3
Mako==1.3.3
mamba @ file:///home/conda/feedstock_root/build_artifacts/mamba-split_1711394305528/work/mamba
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
menuinst @ file:///home/conda/feedstock_root/build_artifacts/menuinst_1705068268702/work
mkl==2024.0.0
mkl-fft==1.3.1
mkl-service==2.4.0
ml_collections==0.1.1
mpmath==1.2.1
msgpack==1.0.8
multidict==6.0.4
multiprocess==0.70.15
munkres==1.1.4
mysql-connector-python==8.1.0
networkx==3.1
ninja==1.11.1
nltk==3.8.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.3.52
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.0.6
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
parso==0.8.4
PasteDeploy==3.1.0
patsy==0.5.6
pbkdf2==1.3
peft==0.4.0
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.11.17
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1706713388748/work
pluggy @ file:///home/conda/feedstock_root/build_artifacts/pluggy_1706116770704/work
plumbum==1.8.2
pooch==1.8.0
portalocker==2.7.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py3Dmol==2.1.0
pyarrow==14.0.2
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat @ file:///home/conda/feedstock_root/build_artifacts/pycosat_1696355775111/work
pycparser==2.21
pydot==2.0.0
Pygments==2.17.2
pyparsing==3.0.9
pypi==2.1
pyramid==2.0.2
pyramid-mailer==0.15.1
pyrosetta @ https://west.rosettacommons.org/pyrosetta/release/release/PyRosetta4.Release.python38.ubuntu.wheel/pyrosetta-2024.19+release.a34b73c40f-cp38-cp38-linux_x86_64.whl#sha256=9a3721c44f139a95c9a060fa31fbc707f39556c932d5c1570068f2100db78f6c
pyrosetta-installer==0.1.1
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1661604839144/work
python-dateutil==2.8.2
python3-openid==3.2.0
pytorch-lightning==1.9.3
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
repoze.sendmail==4.4.1
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
requests-oauthlib==1.3.1
responses==0.18.0
rich==13.7.1
rpyc==5.3.1
rsa==4.9
ruamel.yaml @ file:///home/conda/feedstock_root/build_artifacts/ruamel.yaml_1707298136555/work
ruamel.yaml.clib @ file:///home/conda/feedstock_root/build_artifacts/ruamel.yaml.clib_1707314498139/work
sacrebleu==2.3.1
sacremoses==0.1.1
safetensors==0.4.2
scikit-learn==1.3.0
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==2.1.1
setproctitle==1.3.3
shtab==1.7.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==1.4.52
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
tabulate==0.9.0
tbb==2021.11.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
threadpoolctl==3.2.0
tokenizers==0.19.1
torch==1.13.1+cu117
torchaudio==0.13.1+cu117
torchdiffeq==0.2.3
torchmetrics==1.2.0
torchvision==0.14.1+cu117
tqdm==4.66.1
traitlets==5.14.3
transaction==4.0
transformers==4.41.2
translationstring==1.4
triton==2.3.1
trl==0.8.6
typing_extensions==4.12.1
tyro==0.8.4
tzdata==2023.3
urllib3==2.0.3
velruse==1.1.1
venusian==3.1.0
virtualenv==20.25.0
wandb==0.17.0
wcwidth==0.2.13
WebOb==1.8.7
Werkzeug==3.0.0
win-inet-pton==1.1.0
wincertstore==0.2
WTForms==3.1.1
wtforms-recaptcha==0.3.2
XlsxWriter==3.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.16.2
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1
zstandard==0.22.0
